Bug fix on ssh-select-key and git conf added.
.bash_functions:
- ssh-select-key: Verifies if id_rsa and id_rsa.pub are actual symlinks
  before removing them.

.gitconfig:
- Added push conf "current".

+++ .bash_functions
@@ -111,8 +111,13 @@ function ssh-select-key() {
     fi
     note "Selected key: ${selected}"
     if [ -e ${sshdir}/id_rsa ] &&
-       [ -e ${sshdir}/id_rsa.pub ]; then
+       [ -e ${sshdir}/id_rsa.pub ] &&
+       [ -L ${sshdir}/id_rsa ] &&
+       [ -L ${sshdir}/id_rsa.pub ]; then
         rm ${sshdir}/id_rsa{,.pub}
+    else
+        err "id_rsa files exist and are NOT symlinks!"
+        return 1
     fi
     link ${sshdir}/"${selected}" ${sshdir}/id_rsa
     link ${sshdir}/"${selected}".pub ${sshdir}/id_rsa.pub
